+++ domain/src/main/kotlin/gr/eduinvoice/domain/user/UserUseCases.kt
@@ -6,5 +6,6 @@ data class UserUseCases @Inject constructor(
     val createUser: CreateUser,
     val authenticateUser: AuthenticateUser,
     val getUserProfile: GetUserProfile,
-    val updateUser: UpdateUser
+    val updateUser: UpdateUser,
+    val resetPassword: ResetPassword
 )